Splited translator helper class into two classes
